pyproject.toml
[project]
name = "tweet2embed"
version = "0.1.0"
description = "Convert a public Tweet into either an image with alt text or Semantic HTML and CSS"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "Terence Eden" }]
dependencies = [
    "pillow==10.4.0",
    "pyperclip==1.9.0",
    "requests==2.32.3",
    "click>=8.1.7",
    "python-dateutil>=2.9.0.post0",
    "jinja2>=3.1.4",
]

[project.optional-dependencies]
img = ["selenium==4.23.1", "webdriver-manager==4.0.2"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
